pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "neurobik"
version = "0.1.0"
description = "CLI tool for downloading AI models and OCI images"
readme = "README.md"
license = {text = "GPL-3.0"}
requires-python = ">=3.8"
authors = [
    {name = "Codenomad"},
]
dependencies = [
    "requests>=2.28.0",
    "tqdm>=4.64.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "textual>=0.50.0",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
    "questionary>=2.0.0",
]

[project.scripts]
neurobik = "neurobik.cli:download"

[tool.hatch.build.targets.wheel]
packages = ["neurobik"]